Add stub files for paint functions
