+++ src/OpenLoco/CMakeLists.txt
@@ -245,16 +245,21 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Objects/WaterObject.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/OpenLoco.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/Paint.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintAirport.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintBuilding.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintDocks.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintEffectEntity.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintEntity.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintIndustry.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintRoad.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintRoadStation.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintSignal.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintStation.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintSurface.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintTile.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintTileDecorations.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintTrack.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintTrainStation.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintTree.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintVehicle.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintWall.cpp"
@@ -597,6 +602,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintEffectEntity.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintEntity.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintIndustry.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintRoad.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintSignal.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintStation.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Paint/PaintSurface.h"
